Move each SafeHandle to its own file.
